+++ packages/toolboxes/evm/src/api/covalentApi.ts
@@ -1,4 +1,4 @@
-import { AssetValue, formatBigIntToSafeValue, getRequest } from '@swapkit/helpers';
+import { AssetValue, filterAssets, formatBigIntToSafeValue, getRequest } from '@swapkit/helpers';
 import type { ChainId } from '@swapkit/types';
 import { ChainIdToChain } from '@swapkit/types';
 
@@ -27,13 +27,13 @@ type CovalentBalanceResponse = {
 };
 
 export const covalentApi = ({ apiKey, chainId }: { apiKey: string; chainId: ChainId }) => ({
-  getBalance: async (address: string) => {
+  getBalance: async (address: string, potentialScamFilter?: boolean) => {
     const { data } = await getRequest<{ data: CovalentBalanceResponse }>(
       `https://api.covalenthq.com/v1/${chainId}/address/${address}/balances_v2/`,
       { key: apiKey },
     );
 
-    return (data?.items || []).map(
+    const balances = (data?.items || []).map(
       ({ balance, contract_decimals, contract_ticker_symbol, contract_address }) =>
         new AssetValue({
           value: formatBigIntToSafeValue({ value: BigInt(balance), decimal: contract_decimals }),
@@ -41,6 +41,8 @@ export const covalentApi = ({ apiKey, chainId }: { apiKey: string; chainId: Chai
           identifier: `${ChainIdToChain[chainId]}.${contract_ticker_symbol}-${contract_address}`,
         }),
     );
+
+    return potentialScamFilter ? filterAssets(balances) : balances;
   },
 });