+++ packages/swapkit/helpers/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @swapkit/helpers
 
+## 1.0.0-rc.6
+
+### Patch Changes
+
+- add scam filter for balances
+
+- Updated dependencies []:
+  - @swapkit/tokens@1.0.0-rc.5
+  - @swapkit/types@1.0.0-rc.5
+
 ## 1.0.0-rc.5
 
 ### Patch Changes