+++ packages/swapkit/types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @swapkit/types
 
+## 1.0.0-rc.5
+
+### Patch Changes
+
+- add scam filter for balances
+
 ## 1.0.0-rc.4
 
 ### Patch Changes