+++ packages/toolboxes/utxo/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @swapkit/toolbox-utxo
 
+## 1.0.0-rc.7
+
+### Patch Changes
+
+- add scam filter for balances
+
+- Updated dependencies []:
+  - @swapkit/helpers@1.0.0-rc.6
+  - @swapkit/types@1.0.0-rc.5
+
 ## 1.0.0-rc.6
 
 ### Patch Changes