+++ packages/wallets/xdefi/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @swapkit/wallet-xdefi
 
+## 1.0.0-rc.10
+
+### Patch Changes
+
+- add scam filter for balances
+
+- Updated dependencies []:
+  - @swapkit/toolbox-cosmos@1.0.0-rc.8
+  - @swapkit/helpers@1.0.0-rc.6
+  - @swapkit/toolbox-utxo@1.0.0-rc.7
+  - @swapkit/types@1.0.0-rc.5
+  - @swapkit/toolbox-evm@1.0.0-rc.7
+
 ## 1.0.0-rc.9
 
 ### Patch Changes